Move TCLab to examples
